modificaciones multimedia

+++ src/main/java/com/ubb/proyecto/service/MultimediaService.java
@@ -14,8 +14,16 @@ public class MultimediaService {
     @Autowired
     private RepositorioMultimedia repositorioMultimedia;
 
+    public MultimediaService(RepositorioMultimedia repositorioMultimedia) {
+        this.repositorioMultimedia = repositorioMultimedia;
+    }
+
     public List<Multimedia> getAllMultimedia() {
         return repositorioMultimedia.findAll();
+    }
+
+    public Optional<Multimedia> getMultimediaById(Integer id) {
+        return repositorioMultimedia.findById(id);
     }
 
     public Multimedia addMultimedia(Multimedia multimedia) {
@@ -40,9 +48,17 @@ public class MultimediaService {
         }
     }
 
-    public void deleteMultimedia(int id) {
-        repositorioMultimedia.deleteById(id);
+    public boolean deleteMultimedia(Integer id) {
+        if (repositorioMultimedia.existsById(id)) {
+            repositorioMultimedia.deleteById(id);
+            return true;
+        }
+        return false;
     }
+    
+    /*public void deleteMultimedia(int id) {
+        repositorioMultimedia.deleteById(id);
+    }*/
 
     /*public boolean isValidURL(String url) {
         try {
